+++ Урок_5/L5c1.py
@@ -5,7 +5,7 @@
 строка.
 '''
 
-#Запись с помощью Print()
+#Запись с помощью Print() # Работает, но первая строка остается пустой - некрасиво.
 # with open('Пример.txt', 'a', encoding='utf-8') as f:
 #     word = input('Введите слово или фразу:')
 #     print(word, file=f) #Почему когда используешь функцию Print() первая строка остается пустой?
