+++ Урок_5/L5c2.py
@@ -1,5 +1,10 @@
 # -*- coding: utf8 -*-
 
+'''
+Создать текстовый файл (не программно), сохранить в нём несколько строк, выполнить
+подсчёт строк и слов в каждой строке.
+'''
+
 my_f = open( "Пример.txt", "r" )
 content = my_f.readlines()
 print(f'Количество строк в документе: {len(content)}')
